backend/requirements.txt
asgiref==3.11.1
config==0.5.1
dj-database-url==3.1.2
Django==6.0.2
django-cors-headers==4.9.0
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
numpy==2.4.2
pandas==3.0.1
psycopg==3.3.3
psycopg-binary==3.3.3
PyJWT==2.11.0
python-dateutil==2.9.0.post0
python-decouple==3.8
python-dotenv==1.2.1
six==1.17.0
sqlparse==0.5.5
typing_extensions==4.15.0
